qtquick2app-rust: drop qmake use, simplify rust_env.inc
For now, this is still considered to "use qmake as builder" and installs
template.pro... but it isn't used in the phases in the RPM specfile.  Ideally
the qmake bits would be properly removed and also this should all be tested
against qtcreator, but for now this works well enough with modified sdk-init.